# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/games": {
            "post": {
                "description": "create game",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "parameters": [
                    {
                        "description": "CreateGameInput",
                        "name": "CreateGameInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DTO.CreateGameInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse-DTO_GameResponseOnlyDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/games/": {
            "get": {
                "description": "get all game",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse-DTO_GamesResponseDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "create game preview",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "parameters": [
                    {
                        "description": "CreatePreviewInput",
                        "name": "CreatePreviewInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DTO.CreatePreviewInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse-DTO_PreviewResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/games/preview/": {
            "delete": {
                "description": "delete preview by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Preview ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/games/{id}": {
            "get": {
                "description": "get game by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse-DTO_GameResponseDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "delete game by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "patch": {
                "description": "update game",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "UpdateGameInput",
                        "name": "UpdateGameInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DTO.UpdateGameInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "DTO.CreateGameInput": {
            "type": "object",
            "required": [
                "description",
                "icon",
                "src",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "src": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "DTO.CreatePreviewInput": {
            "type": "object",
            "required": [
                "game_id",
                "image"
            ],
            "properties": {
                "game_id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "video": {
                    "type": "string"
                }
            }
        },
...
        "DTO.GameResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "previews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/DTO.PreviewResponse"
                    }
                },
                "src": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "DTO.GameResponseDTO": {
            "type": "object",
            "properties": {
                "game": {
                    "$ref": "#/definitions/DTO.GameResponse"
                }
            }
        },
        "DTO.GameResponseOnly": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "src": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "DTO.GameResponseOnlyDTO": {
            "type": "object",
            "properties": {
                "game": {
                    "$ref": "#/definitions/DTO.GameResponseOnly"
                }
            }
        },
        "DTO.GamesResponseDTO": {
            "type": "object",
            "properties": {
                "games": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/DTO.GameResponseOnly"
                    }
                }
            }
        },
        "DTO.PreviewResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "video": {
                    "type": "string"
                }
            }
        },
        "DTO.PreviewResponseDTO": {
            "type": "object",
            "properties": {
                "preview": {
                    "$ref": "#/definitions/DTO.PreviewResponse"
                }
            }
        },
...
        "DTO.UpdateGameInput": {
            "type": "object",
            "required": [
                "description",
                "icon",
                "src",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "src": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "api.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "parameter": {
                    "type": "string"
                }
            }
        },
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Error"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "api.SuccessResponse-DTO_GameResponseDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/DTO.GameResponseDTO"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "api.SuccessResponse-DTO_GameResponseOnlyDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/DTO.GameResponseOnlyDTO"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "api.SuccessResponse-DTO_GamesResponseDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/DTO.GamesResponseDTO"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "api.SuccessResponse-DTO_PreviewResponseDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/DTO.PreviewResponseDTO"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}